Preparing android
