+++ android/app/build.gradle
@@ -128,8 +128,6 @@ android {
 dependencies {
     compile project(':react-native-device-info')
     compile project(':react-native-randombytes')
-    compile project(':react-native-cookies')
-    compile project(':react-native-fetch-blob')
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
