+++ js/site_manager.js
@@ -1,0 +1,228 @@
+/**
+ * @flow
+ */
+
+import React, {
+  Component,
+  PropTypes
+} from 'react';
+
+import {
+  AsyncStorage,
+  Platform,
+  PushNotificationIOS
+} from 'react-native';
+
+import Site from './site';
+import RSAKeyPair from 'keypair';
+import DeviceInfo from 'react-native-device-info';
+import randomBytes from 'react-native-randombytes';
+import JSEncrypt from './../lib/jsencrypt';
+
+class SiteManager {
+  constructor() {
+    this._subscribers = [];
+    this.sites = [];
+    this.load();
+    this.ensureRSAKeys();
+  }
+
+  add(site) {
+    this.sites.push(site);
+    this.save();
+    this._onChange();
+  }
+
+  remove(site) {
+    let index = this.sites.indexOf(site);
+    if (index >= 0) {
+      this.sites.splice(index,1);
+      this.save();
+      this._onChange()
+    }
+  }
+
+  subscribe(callback) {
+    this._subscribers.push(callback);
+  }
+
+  unsubscribe(callback) {
+    var pos = this._subscribers.indexOf(callback);
+    if (pos >= -1) {
+      this._subscribers = this._subscribers.splice(pos,1);
+    }
+  }
+
+  save() {
+    AsyncStorage.setItem('@Discourse.sites', JSON.stringify(this.sites));
+
+    if(Platform.OS === 'ios') {
+      PushNotificationIOS.checkPermissions(p => {
+        if (p.badge) {
+          PushNotificationIOS.setApplicationIconBadgeNumber(this.totalUnread());
+        }
+      });
+    }
+  }
+
+  ensureRSAKeys() {
+    AsyncStorage.getItem('@Discourse.rsaKeys').then((json) => {
+      if (json) {
+        this.rsaKeys = JSON.parse(json);
+      } else {
+        this.rsaKeys = RSAKeyPair();
+        AsyncStorage.setItem('@Discourse.rsaKeys', JSON.stringify(this.rsaKeys));
+      }
+    });
+  }
+
+  load() {
+    AsyncStorage.getItem('@Discourse.sites').then((json) => {
+      if (json) {
+        this.sites = JSON.parse(json).map(obj=>new Site(obj));
+        this._onChange()
+      }
+    });
+  }
+
+  totalUnread() {
+    let count = 0;
+    this.sites.forEach((site)=>{
+      if (site.authToken) {
+        count += (site.unreadNotifications || 0) + (site.unreadPrivateMessages || 0);
+      }
+    });
+    return count;
+  }
+
+  refreshSites(opts) {
+    let sites = this.sites.slice(0);
+    opts = opts || {};
+
+    return new Promise((resolve,reject)=>{
+      if (sites.length === 0) {
+        resolve(false);
+        return;
+      }
+
+      let site = sites.pop();
+      let somethingChanged = false;
+
+      let processSite = (site) => {
+        site.refreshNotificationCounts(opts).then((changed) => {
+          somethingChanged = somethingChanged || changed;
+          let s = sites.pop();
+          if (s) { processSite(s); }
+          else {
+            if (somethingChanged) {
+              this.save();
+            }
+            resolve(somethingChanged);
+          }
+        });
+      };
+
+      processSite(site);
+    });
+  }
+
+
+  serializeParams(obj) {
+    return Object.keys(obj)
+                 .map(k => `${encodeURIComponent(k)}=${encodeURIComponent([obj[k]])}`)
+                .join("&");
+  }
+
+  registerClientId(id) {
+    this.getClientId().then(existing => {
+      if (existing !== id) {
+        this.clientId = id;
+        AsyncStorage.setItem('@ClientId', this.clientId);
+        this.sites.forEach((site)=>{
+          site.authToken = null;
+          site.userId = null;
+        });
+        this.save();
+      }
+    });
+  }
+
+  getClientId() {
+    return new Promise(resolve=>{
+      if (this.clientId) {
+        resolve(this.clientId);
+      } else {
+        AsyncStorage.getItem('@ClientId').then((clientId)=>{
+          if(clientId && clientId.length > 0) {
+            this.clientId = clientId;
+            resolve(clientId);
+          } else {
+            randomBytes(32, (err, bytes) => {
+              this.clientId = bytes.toString('hex');
+              AsyncStorage.setItem('@ClientId', this.clientId);
+              resolve(this.clientId);
+            });
+          }
+        });
+      }
+    });
+  }
+
+  generateNonce(site) {
+    return new Promise(resolve=>{
+      randomBytes(16, (err, bytes) => {
+        this._nonce = bytes.toString('hex');
+        this._nonceSite = site;
+        resolve(this._nonce);
+      });
+    });
+  }
+
+  handleAuthPayload(payload) {
+    let crypt = new JSEncrypt();
+
+    crypt.setKey(this.rsaKeys.private);
+    let decrypted = JSON.parse(crypt.decrypt(payload));
+
+    if (decrypted.nonce !== this._nonce) {
+      alert("We were not expecting this reply, please try again!");
+      return;
+    }
+
+    this._nonceSite.authToken = decrypted.key;
+    this._nonceSite.refreshNotificationCounts()
+        .then(()=>{
+          this.save();
+          this._onChange();
+        });
+  }
+
+  generateAuthURL(site) {
+    let clientId;
+
+    return this.getClientId()
+      .then(cid => {
+        clientId = cid;
+        return this.generateNonce(site);
+      })
+      .then(nonce => {
+         let params = {
+          access: 'rp',
+          client_id: clientId,
+          nonce: nonce,
+          push_url: 'https://api.discourse.org/api/ios_notify',
+          auth_redirect: 'discourse://auth_redirect',
+          application_name: "Discourse - " + DeviceInfo.getDeviceName(),
+          public_key: this.rsaKeys.public
+        };
+
+        return `${site.url}/user-api-key/new?${this.serializeParams(params)}`;
+      });
+  }
+
+  _onChange() {
+    this._subscribers.forEach((sub) => sub());
+  }
+}
+
+export default SiteManager;
