+++ package.json
@@ -5,17 +5,18 @@
   "scripts": {
     "start": "node node_modules/react-native/local-cli/cli.js start",
     "storybook": "storybook start -p 9001",
-    "storybook-ios": "storybook run-ios -a DiscourseMobileStorybook -i org.reactjs.native.example.DiscourseMobileStorybook"
+    "storybook-ios": "storybook run-ios -a DiscourseMobileStorybook -i org.discourse.DiscourseMobileStorybook",
+    "clean-ios": "cd ios && xcodebuild -scheme DiscourseMobile -destination 'platform=iOS Simulator,name=iPhone 6' clean"
   },
   "dependencies": {
     "keypair": "^1.0.0",
     "react": "15.2.1",
     "react-native": "^0.31",
     "react-native-device-info": "^0.9.4",
-    "react-native-fetch-blob": "^0.8.2",
     "react-native-randombytes": "^1.0.1",
     "react-native-safari-view": "^1.0.1",
-    "react-native-swipeout": "git+https://github.com/dancormier/react-native-swipeout.git"
+    "react-native-swipeout": "git+https://github.com/dancormier/react-native-swipeout.git",
+    "react-native-ecc": "^1.2.2"
   },
   "devDependencies": {
     "@kadira/react-native-storybook": "^1.12.3"
